rag/requirements.txt
fastapi
uvicorn
boto3
pandas
cassandra-driver
sentence-transformers
numpy
faiss-cpu
